Backend loginpage

+++ Frontend/components/subtask/userIconDelete.jsx
@@ -30,8 +30,19 @@ function Delete({user}) {
         console.log(user.user_id)
         console.log(subtask_id)
 
+        const accessToken = sessionStorage.getItem('accessToken');
+        //Redirect to login if there's no access token
+        if (!accessToken) {
+            window.location.href = "http://localhost:5173/login"
+            return;
+        }
+
         fetch(`http://localhost:8000/user-subtasks/${user.user_id}/${parseInt(subtask_id)}/`, {
             method: 'DELETE',
+            headers: {
+                'Authorization': `JWT ${accessToken}`, 
+                'Content-Type': 'application/json',
+            },
         })
         .then(() => {
             console.log('Deleted')
